gallery page and order page
gallery page and order page changed from flex box to grid format when it comes to the lipgloss photos .

also fixed the issue where going from gallery page to home page, it gave an error because of the path being incorrect on the anchor. was set to "homepage.html" instead of "index.html".